Update sync.yml
Using an alternative action ([dabreadman/sync-upstream-repo@v1.3.0](https://github.com/marketplace/actions/sync-and-merge-upstream-repository-with-your-current-repository)), I was able to update the fork, see the results [here](https://github.com/PhiSpel/Modellbildung-und-Simulation-1/actions/runs/3839309334).

Updating the `gh-pages` branch throws an issue merging `README.md` (which are identical across all three forks), but marks the workflow run as successful. So I suppose, it is not that big a deal.